-refactor error in crud operation

+++ delete.go
@@ -5,41 +5,38 @@ import (
 )
 
 // DeleteObjects borra objetos de la base de datos según nombre de la tabla y ids.
-func (c Connection) DeleteObjects(table_name string, all_data ...map[string]string) (message string, ok bool) {
+func (c Connection) DeleteObjects(table_name string, all_data ...*map[string]string) error {
 	c.Open()
 	defer c.Close()
 
 	tx, err := c.Begin()
 	if err != nil {
-		filterMessageDBtoClient(err.Error(), table_name, &message)
-		return
+		return filterMessageDBtoClient(err, table_name)
 	}
 
 	for _, data := range all_data {
+		if data != nil {
+			// borramos el objeto usando la clave primaria como condición
+			query := fmt.Sprintf("DELETE FROM %s WHERE %s = %s", table_name, "id_"+table_name, c.PlaceHolders(1))
+			stmt, err := tx.Prepare(query)
+			if err != nil {
+				tx.Rollback()
+				return filterMessageDBtoClient(err, table_name, data)
+			}
+			defer stmt.Close()
 
-		// borramos el objeto usando la clave primaria como condición
-		query := fmt.Sprintf("DELETE FROM %s WHERE %s = %s", table_name, "id_"+table_name, c.PlaceHolders(1))
-		stmt, err := tx.Prepare(query)
-		if err != nil {
-			filterMessageDBtoClient(err.Error(), table_name, &message, data)
-			tx.Rollback()
-			return
-		}
-		defer stmt.Close()
-
-		_, err = stmt.Exec(data["id_"+table_name])
-		if err != nil {
-			filterMessageDBtoClient(err.Error(), table_name, &message, data)
-			tx.Rollback()
-			return
+			_, err = stmt.Exec((*data)["id_"+table_name])
+			if err != nil {
+				tx.Rollback()
+				return filterMessageDBtoClient(err, table_name, data)
+			}
 		}
 	}
 
 	if err := tx.Commit(); err != nil {
-		filterMessageDBtoClient(err.Error(), table_name, &message)
 		tx.Rollback()
-		return
+		return filterMessageDBtoClient(err, table_name)
 	}
 
-	return "Registro(s) Eliminado(s)", true
+	return nil
 }
